+++ test/js/base/api.js
@@ -5,6 +5,55 @@ const ripe = require("../../../src/js");
 describe("RipeAPI", function() {
     this.timeout(config.TEST_TIMEOUT);
 
+    describe("#_authCallback", function() {
+        it("should be able to retry operations", async () => {
+            const remote = ripe.RipeAPI();
+            let options = {
+                retries: 2,
+                dummy: 3,
+                authCallback: callback => {
+                    options.dummy--;
+                    if (callback) callback();
+                }
+            };
+            await new Promise((resolve, reject) => {
+                remote._cacheURL("https://httpbin.org/status/403", options, () => resolve());
+            });
+            assert.strictEqual(options.retries, 0);
+            assert.strictEqual(options.dummy, 1);
+
+            options = {
+                dummy: 3,
+                authCallback: callback => {
+                    options.dummy--;
+                    if (callback) callback();
+                }
+            };
+            await new Promise((resolve, reject) => {
+                remote._cacheURL("https://httpbin.org/status/403", options, () => resolve());
+            });
+            assert.strictEqual(options.retries, 0);
+            assert.strictEqual(options.dummy, 2);
+        });
+
+        it("should not retry operations, if not required", async () => {
+            const remote = ripe.RipeAPI();
+            const options = {
+                retries: 2,
+                dummy: 3,
+                authCallback: callback => {
+                    options.dummy--;
+                    if (callback) callback();
+                }
+            };
+            await new Promise((resolve, reject) => {
+                remote._cacheURL("https://httpbin.org/status/200", options, () => resolve());
+            });
+            assert.strictEqual(options.retries, 2);
+            assert.strictEqual(options.dummy, 3);
+        });
+    });
+
     describe("#_queryToSpec()", function() {
         it("should be able to convert a query to spec", async () => {
             const remote = ripe.RipeAPI();