+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Support for multiple character escaping in `escape()`
 * Add method to retrieve the url for an attachment
+* Support for retry in API requests when receiving authentication related errors (eg: 403)
 
 ### Changed